refactor: Validator view단에서 분리, WinningLotto 검증 로직 추가
