+++ src/main/java/domain/WinningLotto.java
@@ -2,11 +2,15 @@ package domain;
 
 import java.util.List;
 
+import static validation.LottoValidator.*;
+
 public class WinningLotto {
     private final List<Integer> winningNumbers;
     private final int bonusBall;
 
     public WinningLotto(List<Integer> winningNumbers, int bonusBall) {
+        validateWinningNumbers(winningNumbers);
+        validateBonusBallUnique(winningNumbers, bonusBall);
         this.winningNumbers = winningNumbers;
         this.bonusBall = bonusBall;
     }
